switch to CFRunLoopTimerCreateWithHandler for more accurate timing

+++ BareMetal/DrawOperation.swift
@@ -8,7 +8,7 @@
 
 protocol DrawOperation {
     var type: String { get }
-    var timestamp: Int64 { get }
+    var timestamp: Double { get }
     var id: Int64 { get }
 }
 
@@ -16,7 +16,7 @@ struct Line: DrawOperation {
     var type: String
     var start: [Float]
     var end: [Float]
-    var timestamp: Int64 = 0
+    var timestamp: Double = 0
     var id: Int64
 
     init(start: [Float], end: [Float], timestamp _: Int64, id: Int64) {
@@ -34,10 +34,10 @@ struct CubicBezier: DrawOperation {
     var control1: [Float]
     var control2: [Float]
     var lineWidth: Float = 0.050
-    var timestamp: Int64 = 0
+    var timestamp: Double = 0
     var id: Int64
 
-    init(start: [Float], end: [Float], control1: [Float], control2: [Float], timestamp: Int64, id: Int64) {
+    init(start: [Float], end: [Float], control1: [Float], control2: [Float], timestamp: Double, id: Int64) {
         self.type = "CubicBezier"
         self.start = start
         self.end = end
@@ -53,10 +53,10 @@ struct QuadraticBezier: DrawOperation {
     var start: [Float]
     var end: [Float]
     var control: [Float]
-    var timestamp: Int64 = 0
+    var timestamp: Double = 0
     var id: Int64
 
-    init(start: [Float], end: [Float], control: [Float], timestamp: Int64, id: Int64) {
+    init(start: [Float], end: [Float], control: [Float], timestamp: Double, id: Int64) {
         self.type = "QuadraticBezier"
         self.start = start
         self.end = end
@@ -69,10 +69,10 @@ struct QuadraticBezier: DrawOperation {
 struct Pan: DrawOperation {
     var type: String
     var point: [Float]
-    var timestamp: Int64
+    var timestamp: Double
     var id: Int64
 
-    init(point: [Float], timestamp: Int64, id: Int64) {
+    init(point: [Float], timestamp: Double, id: Int64) {
         self.type = "Pan"
         self.point = point
         self.timestamp = timestamp
@@ -83,10 +83,10 @@ struct Pan: DrawOperation {
 struct Point: DrawOperation {
     var type: String
     var point: [Float]
-    var timestamp: Int64
+    var timestamp: Double
     var id: Int64
 
-    init(point: [Float], timestamp: Int64, id: Int64) {
+    init(point: [Float], timestamp: Double, id: Int64) {
         self.type = "Point"
         self.point = point
         self.timestamp = timestamp
@@ -98,10 +98,10 @@ struct PenDown: DrawOperation {
     var type: String
     var color: [Float]
     var lineWidth: Float
-    var timestamp: Int64
+    var timestamp: Double
     var id: Int64
 
-    init(color: [Float], lineWidth: Float, timestamp: Int64, id: Int64) {
+    init(color: [Float], lineWidth: Float, timestamp: Double, id: Int64) {
         self.type = "PenDown"
         self.color = color
         self.lineWidth = lineWidth
@@ -112,10 +112,10 @@ struct PenDown: DrawOperation {
 
 struct PenUp: DrawOperation {
     var type: String
-    var timestamp: Int64 = 0
+    var timestamp: Double = 0
     var id: Int64
 
-    init(timestamp: Int64, id: Int64) {
+    init(timestamp: Double, id: Int64) {
         self.type = "PenUp"
         self.timestamp = timestamp
         self.id = id
